+++ Packages/com.kybernetik.ultevents/Runtime/Misc/Benchmarks/SimpleTimer.cs
@@ -1,4 +1,4 @@
-﻿// UltEvents // https://kybernetik.com.au/ultevents // Copyright 2021-2024 Kybernetik //
+﻿// UltEvents // https://kybernetik.com.au/ultevents // Copyright 2021-2025 Kybernetik //
 
 using System;
 using System.Diagnostics;