+++ Packages/com.kybernetik.ultevents/Runtime/Misc/Operators.cs
@@ -1,4 +1,4 @@
-﻿// UltEvents // https://kybernetik.com.au/ultevents // Copyright 2021-2024 Kybernetik //
+﻿// UltEvents // https://kybernetik.com.au/ultevents // Copyright 2021-2025 Kybernetik //
 
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member.